optimise: remount read-only store writable
The GC binary already remounts /nix/store rw, but fast-nix-optimise did
not, causing EROFS when hardlinking into .links/ (issue #7).

Move make_store_writable into common and call it from both.
We now also enter a private mount namespace so the rw remount stays scoped to the process.

Fixes #7
