+++ crates/common/Cargo.toml
@@ -8,3 +8,5 @@ anyhow.workspace = true
 rusqlite.workspace = true
 foldhash.workspace = true
 harmonia-store-core.workspace = true
+nix.workspace = true
+log.workspace = true
