+++ crates/gc/src/lib.rs
@@ -2,4 +2,4 @@ pub mod gc;
 pub mod profiles;
 pub mod roots;
 
-pub use fast_nix_common::{HashMap, HashSet, db, format_size};
+pub use fast_nix_common::{HashMap, HashSet, db, format_size, make_store_writable};
